Reload the plugins table when a new plugin is installed
